Only move build artifacts if they exist, add deb artifacts
Previously the rake taks would only try to move rpms into the DESTDIR, but it
would do it even if there were no rpms to move. This commit adds debs to the
build artifacts to move and only moves them if there are any. It also moves
the moving of artifacts into its own method.